validation,security

+++ carnet_sanatate/src/Controller/AnimalController.php
@@ -20,6 +20,8 @@ class AnimalController extends AbstractController
      */
     public function index(AnimalRepository $animalRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         return $this->render('animal/index.html.twig', [
             'animals' => $animalRepository->findAll(),
         ]);
@@ -30,6 +32,8 @@ class AnimalController extends AbstractController
      */
     public function new(Request $request): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         $animal = new Animal();
         $form = $this->createForm(Animal3Type::class, $animal);
         $form->handleRequest($request);
@@ -39,7 +43,7 @@ class AnimalController extends AbstractController
             $entityManager->persist($animal);
             $entityManager->flush();
 
-            return $this->redirectToRoute('animal_index');
+            return $this->redirectToRoute('user_animal_new');
         }
 
         return $this->render('animal/new.html.twig', [
@@ -53,6 +57,8 @@ class AnimalController extends AbstractController
      */
     public function show(Animal $animal): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         return $this->render('animal/show.html.twig', [
             'animal' => $animal,
         ]);
@@ -63,13 +69,16 @@ class AnimalController extends AbstractController
      */
     public function edit(Request $request, Animal $animal): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         $form = $this->createForm(Animal3Type::class, $animal);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('animal_index');
+
+            return $this->redirectToRoute('dashboard');
         }
 
         return $this->render('animal/edit.html.twig', [
@@ -83,12 +92,14 @@ class AnimalController extends AbstractController
      */
     public function delete(Request $request, Animal $animal): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         if ($this->isCsrfTokenValid('delete'.$animal->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($animal);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('animal_index');
+        return $this->redirectToRoute('dashboard');
     }
 }
